+++ src/main/java/com/amadeus/Shopping.java
@@ -2,7 +2,7 @@ package com.amadeus;
 
 import com.amadeus.shopping.Activities;
 import com.amadeus.shopping.Activity;
-import com.amadeus.shopping.FlightAvailabilitiesSearch;
+import com.amadeus.shopping.availability.FlightAvailabilities;
 import com.amadeus.shopping.FlightDates;
 import com.amadeus.shopping.FlightDestinations;
 import com.amadeus.shopping.FlightOffers;
@@ -69,7 +69,7 @@ public class Shopping {
    * <code>/v1/shopping/availability/flight-availabilities</code> endpoints.
    * </p>
    */
-  public FlightAvailabilitiesSearch flightAvailabilitiesSearch;
+  public FlightAvailabilities flightAvailabilities;
 
   /**
    * <p>
@@ -125,7 +125,7 @@ public class Shopping {
     this.flightOffersSearch = new FlightOffersSearch(client);
     this.seatMaps = new SeatMaps(client);
     this.activities = new Activities(client);
-    this.flightAvailabilitiesSearch = new FlightAvailabilitiesSearch(client);
+    this.flightAvailabilities = new FlightAvailabilities(client);
   }
 
   /**
